+++ wazo_confd/plugins/group_schedule/resource.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_confd.auth import required_acl
@@ -15,18 +15,18 @@ class GroupScheduleItem(ConfdResource):
         self.group_dao = group_dao
         self.schedule_dao = schedule_dao
 
-    @required_acl('confd.groups.{group_id}.schedules.{schedule_id}.delete')
-    def delete(self, group_id, schedule_id):
+    @required_acl('confd.groups.{group_uuid}.schedules.{schedule_id}.delete')
+    def delete(self, group_uuid, schedule_id):
         tenant_uuids = self._build_tenant_list({'recurse': True})
-        group = self.group_dao.get(group_id, tenant_uuids=tenant_uuids)
+        group = self.group_dao.get(group_uuid, tenant_uuids=tenant_uuids)
         schedule = self.schedule_dao.get(schedule_id, tenant_uuids=tenant_uuids)
         self.service.dissociate(group, schedule)
         return '', 204
 
-    @required_acl('confd.groups.{group_id}.schedules.{schedule_id}.update')
-    def put(self, group_id, schedule_id):
+    @required_acl('confd.groups.{group_uuid}.schedules.{schedule_id}.update')
+    def put(self, group_uuid, schedule_id):
         tenant_uuids = self._build_tenant_list({'recurse': True})
-        group = self.group_dao.get(group_id, tenant_uuids=tenant_uuids)
+        group = self.group_dao.get(group_uuid, tenant_uuids=tenant_uuids)
         schedule = self.schedule_dao.get(schedule_id, tenant_uuids=tenant_uuids)
         self.service.associate(group, schedule)
         return '', 204